requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.4.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
frozenlist==1.4.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
magic-filter==1.0.12
multidict==6.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.2
requests==2.31.0
